cast HINSTANCE to int, which is correct since MSDN docs suggest that, it is not a true HINSTANCE
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@39985 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSWorkspace.m
@@ -814,7 +814,7 @@ static NSDictionary		*urlPreferences = nil;
       unichar *buffer = (unichar *)calloc(1, ([fullPath length] + 1) * sizeof(unichar));
       [fullPath getCharacters: buffer range: NSMakeRange(0, [fullPath length])];
       buffer[[fullPath length]] = 0;
-      BOOL success = (ShellExecuteW(GetDesktopWindow(), L"open", buffer, NULL, 
+      BOOL success = ((int)ShellExecuteW(GetDesktopWindow(), L"open", buffer, NULL, 
                                     NULL, SW_SHOWNORMAL) > 32);
       free(buffer);
       return success;
